Revised bundle name detection in AbstractFxmlView.getBundleName()
Previously getBundleName() always returned the value from
FXMLView#bundle attribute, due to testing the annotation instead of
the bundle-name for being empty.

We now correctly test the configured bundle-name. If the bundle name is
null we or empty we derive the bundle name from the classname.
Otherwise we use the value as is.

Fixes #38
Can be backported to 1.4.x